+++ modules/tinymce/src/themes/silver/main/ts/ui/toolbar/CommonToolbar.ts
@@ -34,6 +34,7 @@ export interface ToolbarSpec {
 export interface MoreDrawerToolbarSpec extends ToolbarSpec {
   readonly getSink: () => Result<AlloyComponent, string>;
   readonly moreDrawerData: MoreDrawerData;
+  readonly onToggled: (comp: AlloyComponent, state: boolean) => void;
 }
 
 export interface ToolbarGroup {
@@ -161,7 +162,9 @@ const renderFloatingMoreToolbar = (toolbarSpec: MoreDrawerToolbarSpec): SketchSp
     components: [ primary ],
     markers: {
       overflowToggledClass: ToolbarButtonClasses.Ticked
-    }
+    },
+    onOpened: (comp) => toolbarSpec.onToggled(comp, true),
+    onClosed: (comp) => toolbarSpec.onToggled(comp, false)
   });
 };
 
@@ -195,10 +198,12 @@ const renderSlidingMoreToolbar = (toolbarSpec: MoreDrawerToolbarSpec): SketchSpe
     onOpened: (comp) => {
       // TINY-9223: This will only broadcast to the same mothership as the toolbar
       comp.getSystem().broadcastOn([ Channels.toolbarHeightChange() ], { type: 'opened' });
+      toolbarSpec.onToggled(comp, true);
     },
     onClosed: (comp) => {
       // TINY-9223: This will only broadcast to the same mothership as the toolbar
       comp.getSystem().broadcastOn([ Channels.toolbarHeightChange() ], { type: 'closed' });
+      toolbarSpec.onToggled(comp, false);
     }
   });
 };