+++ modules/alloy/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Added the `onToggled` callback for the `FloatingToolbarButton` component. #TINY-9271
+- Added the `onOpened` and `onClosed` callbacks for the `SplitFloatingToolbar` component. #TINY-9271
+
 ### Removed
 - Moved `TestStore` to Agar. #TINY-9157