+++ modules/alloy/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
+- Added `previousSelector` optional property to `MenuMatrixMovementSpec`, to start with the selected item in focus. #TINY-9283
 - Added the `onToggled` callback for the `FloatingToolbarButton` component. #TINY-9271
 - Added the `onOpened` and `onClosed` callbacks for the `SplitFloatingToolbar` component. #TINY-9271